Change file names so that the docs are reachable
This is probably not quite right for two reasons.  (1) The document is
broader than just the SDK.  (2) I have chosen to use the term file
store rather than object store.  The document could change further to
conform to the TOC but I want to at least raise the possibility that
it's the TOC that should change.