Document programming with "file stores" (aka object stores or blob stores) (#11)
* WIP for object stores document

* More progress on what is now called "file stores"

* Change file names so that the docs are reachable

This is probably not quite right for two reasons.  (1) The document is
broader than just the SDK.  (2) I have chosen to use the term file
store rather than object store.  The document could change further to
conform to the TOC but I want to at least raise the possibility that
it's the TOC that should change.

* Trying again to connect the new document properly

This time I did amend the sidebar contents.  I think the new document
subsumes what was "web SDK" and "object SDK".

* Still figuring out docusaurus

Just trying to get the sidebar to refer to my document properly.

* Still tinkering

* Apply suggestions from code review

Committing many (thought not all) of Rodrics' suggestions.

Co-authored-by: rodric rabbah <rodric@gmail.com>

* WIP on file-stores document

* Finish a reviewable draft of the file-stores doc

Co-authored-by: rodric rabbah <rodric@gmail.com>